requirements.txt
pandas==2.1.0
numpy==1.24.3
openpyxl==3.1.2
xlrd==2.0.1
pyyaml==6.0
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
pathlib==1.0.1
typing-extensions==4.7.1